refactor: #272 - 블록 메뉴의 액션 구성 로직을 UI와 분리
